+++ src/js/desktop/app.js
@@ -2367,4 +2367,20 @@ var isServer = true // eslint-disable-line no-unused-vars
       console.error(err)
     })
   })
+
+  document.querySelector('#settings-section_connection-info_address').addEventListener('click', function (e) {
+    copyTextToClipboard(e.currentTarget.innerHTML)
+    domManipulations.showToast({
+      'innerText': chrome.i18n.getMessage('copiedToClipboard') || 'Copied to clipboard'
+    })
+  })
+
+  function copyTextToClipboard (text) {
+    var copyFrom = document.createElement('textarea')
+    copyFrom.textContent = text
+    document.body.appendChild(copyFrom)
+    copyFrom.select()
+    document.execCommand('copy')
+    document.body.removeChild(copyFrom)
+  }
 }())